I am like 100% sure I did the RSA wrong, but here you go anyone who is competent. Figure out the Redirect opcode, and the handling of the second socket, and you'll have a core set up for non client modification use. I'll probably return to this later, but that fat sexy bitch trunk is greedy and needs more love.

+++ Branch/Hellbound/src/hearthstone-world/WorldSession.h
@@ -268,7 +268,7 @@ public:
 		_recvQueue.Push(packet);
 	}
 
-	void OutPacket(uint16 opcode, uint16 len, const void* data)
+	void OutPacket(uint32 opcode, uint16 len, const void* data)
 	{
 		if(_socket && _socket->IsConnected())
 			_socket->OutPacket(opcode, len, data);
