+++ Branch/Hellbound/src/hearthstone-world/WorldSocket.h
@@ -48,8 +48,8 @@ public:
 	HEARTHSTONE_INLINE void SendPacket(WorldPacket* packet) { if(!packet) return; OutPacket(packet->GetOpcode(), packet->size(), (packet->size() ? (const void*)packet->contents() : NULL)); }
 	HEARTHSTONE_INLINE void SendPacket(StackPacket * packet) { if(!packet) return; OutPacket(packet->GetOpcode(), packet->GetSize(), (packet->GetSize() ? (const void*)packet->GetBufferPointer() : NULL)); }
 
-	void __fastcall OutPacket(uint16 opcode, size_t len, const void* data);
-	OUTPACKET_RESULT __fastcall _OutPacket(uint16 opcode, size_t len, const void* data);
+	void __fastcall OutPacket(uint32 opcode, size_t len, const void* data);
+	OUTPACKET_RESULT __fastcall _OutPacket(uint32 opcode, size_t len, const void* data);
 
 	HEARTHSTONE_INLINE uint32 GetLatency() { return _latency; }
